Do not select windows when updating display tables
Selecting windows during minor mode initialization may have
undesirable effects such as resetting a buffer's point to the
window-point of one of the windows it is displayed in.  This can cause
problems when Lisp code wants to use a temp buffer (whose major mode
happens to enable page-break-lines-mode).

For example, when org-babel-insert-result inserts a table result into
an org buffer, it moves point to the place where the result should be
inserted, then calls orgtbl-to-orgtbl (which uses a temp buffer in
org-mode) to calculate the string to insert, and then inserts that
string directly.  If page-break-lines-mode runs in that temp buffer
and causes the original org buffer's window to be selected, that
buffer's point will move and the results will be inserted in the wrong
place.

Fix #45.